Chat QoL updates Announcement and Event notifications department/group specific

+++ app/src/main/java/com/example/first_responder_app/fragments/ChatFragment.java
@@ -5,6 +5,9 @@ import static android.content.ContentValues.TAG;
 import androidx.databinding.DataBindingUtil;
 import androidx.lifecycle.ViewModelProvider;
 
+import android.app.AlertDialog;
+import android.content.Context;
+import android.content.DialogInterface;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
@@ -21,6 +24,7 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.inputmethod.InputMethodManager;
 
 import com.example.first_responder_app.FirestoreDatabase;
 import com.example.first_responder_app.dataModels.UsersDataModel;
@@ -33,6 +37,7 @@ import com.example.first_responder_app.viewModels.ChatViewModel;
 import com.example.first_responder_app.R;
 import com.google.firebase.Timestamp;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.ListenerRegistration;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 
 import java.util.ArrayList;
@@ -51,6 +56,8 @@ public class ChatFragment extends Fragment {
     private List<String> listOfMembers;
     private ChatRecyclerViewAdapter chatRecyclerViewAdapter;
     private Chat c;
+    private RecyclerView chatRecyclerView;
+    private ListenerRegistration chatListener;
 
     // ref: https://github.com/ScaleDrone/android-chat-tutorial/blob/master/app/src/main/res/layout/activity_main.xml
     public static ChatFragment newInstance() {
@@ -86,17 +93,21 @@ public class ChatFragment extends Fragment {
         mViewModel = new ViewModelProvider(requireActivity()).get(ChatViewModel.class);
         messageInfo = mViewModel.getMessageDetail();
 
-        RecyclerView chatRecyclerView = binding.chatRecyclerView;
+        chatRecyclerView = binding.chatRecyclerView;
         chatRecyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
         chatRecyclerViewAdapter = new ChatRecyclerViewAdapter(getContext(), listOfMessages, listOfMembers);
         //chatRecyclerViewAdapter.setClickListener(chatClickListener);
         chatRecyclerView.setAdapter(chatRecyclerViewAdapter);
 
+
+
+
         chatRecyclerView.addOnLayoutChangeListener(new View.OnLayoutChangeListener() {
             @Override
             public void onLayoutChange(View view, int i, int i1, int i2, int i3, int i4, int i5, int i6, int i7) {
+                Log.d(TAG, "onLayoutChange: scroll " + listOfMessages.size());
                 chatRecyclerView.removeOnLayoutChangeListener(this);
-                chatRecyclerView.scrollToPosition(chatRecyclerView.getAdapter().getItemCount() - 1);
+                chatRecyclerView.scrollToPosition(chatRecyclerViewAdapter.getItemCount() - 1);
             }
         });
 
@@ -108,6 +119,15 @@ public class ChatFragment extends Fragment {
                 userName = user.getFirst_name() + " " + user.getLast_name();
             }
             String msg = binding.editText.getText().toString();
+            binding.editText.setText("");
+
+            //hide keyboard
+            View view = getActivity().getCurrentFocus();
+            if (view != null) {
+                InputMethodManager imm = (InputMethodManager)getActivity().getSystemService(Context.INPUT_METHOD_SERVICE);
+                imm.hideSoftInputFromWindow(view.getWindowToken(), 0);
+            }
+            binding.editText.clearFocus();
 
             if (!msg.equals("") && user != null) {
                 firestoreDatabase.addMessage(c.getId(), msg, user.getDocumentId(), chatRecyclerViewAdapter, mViewModel);
@@ -116,10 +136,18 @@ public class ChatFragment extends Fragment {
         });
 
         binding.leaveChat.setOnClickListener(v -> {
-            firestoreDatabase.removeUserFromChat(user.getDocumentId(), c.getId(), listOfMembers);
-            NavDirections action = ChatFragmentDirections.actionChatFragmentToChatGroupFragment();
-            Navigation.findNavController(binding.getRoot()).navigate(action);
 
+            new AlertDialog.Builder(getActivity())
+                    .setTitle("Leave Chat?")
+                    .setMessage("Are you sure you want to leave this chat?")
+                    .setPositiveButton("Yes", (DialogInterface.OnClickListener) (dialogInterface, i) -> {
+                        firestoreDatabase.removeUserFromChat(user.getDocumentId(), c.getId(), listOfMembers);
+                        NavDirections action = ChatFragmentDirections.actionChatFragmentToChatGroupFragment();
+                        Navigation.findNavController(binding.getRoot()).navigate(action);
+                        firestoreDatabase.addMessage(c.getId(), user.getFull_name() + " has left the chat", user.getDocumentId(), chatRecyclerViewAdapter, mViewModel);
+                    })
+                    .setNegativeButton("No", null)
+                    .show();
         });
 
         return binding.getRoot();
@@ -134,12 +162,13 @@ public class ChatFragment extends Fragment {
 
     private void populateMessageList() {
         String chatId = c.getId();
-        db.collection("chat").document(chatId).collection("messages")
-                .get().addOnCompleteListener(task -> {
-                    if (task.isSuccessful()) {
+        chatListener = db.collection("chat").document(chatId).collection("messages")
+                .addSnapshotListener((snapshot, e) -> {
+                    if (e == null) {
                         ArrayList<Message> temp = new ArrayList<>();
 
-                        for (QueryDocumentSnapshot doc : task.getResult()){
+
+                        for (QueryDocumentSnapshot doc : snapshot){
                             Timestamp t = (Timestamp) doc.get("time_sent");;
                             Message m = new Message(doc.getId(), (String) doc.get("message_text"), (String) doc.get("sender"), t);
                             temp.add(m);
@@ -148,13 +177,20 @@ public class ChatFragment extends Fragment {
                         listOfMessages.addAll(temp);
                         Collections.sort(listOfMessages);
                         mViewModel.setListOfMessages(listOfMessages);
-
+                        chatRecyclerView.scrollToPosition(chatRecyclerViewAdapter.getItemCount() - 1);
                         chatRecyclerViewAdapter.notifyDataSetChanged();
                     } else {
-                        Log.d(TAG, "db get failed in chat page " + task.getException());
+                        Log.d(TAG, "db get failed in chat page " + e);
                     }
                 });
     }
 
 
+    @Override
+    public void onDestroyView() {
+        super.onDestroyView();
+        if(chatListener != null){
+            chatListener.remove();
+        }
+    }
 }
